requirements.txt
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
Cython==0.29.11
EasyProcess==0.2.7
enum34==1.1.6
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
playsound==1.2.2
PyAutoGUI==0.9.45
pycodestyle==2.5.0
pygame==1.9.6
PyGetWindow==0.0.6
pylint==2.3.1
PyMsgBox==1.0.7
pynput==1.4.2
PyRect==0.1.4
pyscreenshot==0.5.1
PyScreeze==0.1.22
pytesseract==0.2.7
python-xlib==0.25
python3-xlib==0.15
PyTweening==1.0.3
six==1.12.0
typed-ast==1.4.0
virtualenv==16.6.1
wrapt==1.11.2
xlib==0.21
Image==1.5.27